% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{available_filters}
\alias{available_filters}
\title{Return a list of installed filters}
\usage{
available_filters()
}
\value{
A character vector containing the names of the installed filters.
}
\description{
This function can be used to produce a vector of the installed filters.
}
\examples{
available_filters()

}
