% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_filters.R
\name{filters}
\alias{filters}
\title{Set HDF5 compression filters}
\arguments{
\item{h5plist}{HDF5 dataset creation property list}

\item{h5tid}{HDF5 type ID}

\item{level}{Compression level to use}

\item{method}{When using the BLOSC filter, defines which compression tool 
should be used.}

\item{shuffle}{Logical value determining whether byte shuffling should be
applied before compression is applied. This only applies to the BLOSC 
filter (see details below).}
}
\description{
Set compressions filters when working with dataset creation property lists.
}
\details{
These functions can be used to set the compression filter when working
with dataset creation property lists directly.
}
