% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{hdf5_plugin_path}
\alias{hdf5_plugin_path}
\title{Return location of installed filters}
\usage{
hdf5_plugin_path()
}
\value{
A character string containing the path where the compiled filters
are located.
}
\description{
This function can be used to find the location of the installed filters.
Setting the environment variable HDF5_PLUGIN_PATH to this value will allow
other applications that require the the filters to use the versions
distributed with this package.
}
\examples{
hdf5_plugin_path()

}
